python/requirements.txt
fastapi==0.115.0
uvicorn[standard]
httpx==0.27.2
pydantic==2.10.0
playwright==1.48.0
pandas==2.2.2
openpyxl==3.1.5
numpy==1.26.4
pdfplumber==0.11.4
PyPDF2==3.0.1
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
tenacity==9.0.0
google-api-python-client==2.152.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

